rename inbox FILES and add them to index
