+++ ts/IterableInbox.tsx
@@ -1,8 +1,8 @@
 'use strict'
 import React, { Component, useState } from 'react'
 import { Text, View, StyleSheet } from 'react-native'
-import MessageList from './MessageList'
-import EmptyState from './EmptyState'
+import MessageList from './IterableInboxMessageList'
+import EmptyState from './IterableInboxEmptyState'
 import sampleMessages from './sampleMessageData.js' 
 
 const Inbox = () => {
